git-debimport audit_2.4-1.debian.tar.xz

+++ src/ausearch-llist.h
@@ -1,6 +1,6 @@
 /*
 * ausearch-llist.h - Header file for ausearch-llist.c
-* Copyright (c) 2005-2008, 2013 Red Hat Inc., Durham, North Carolina.
+* Copyright (c) 2005-2008, 2013-14 Red Hat Inc., Durham, North Carolina.
 * Copyright (c) 2011 IBM Corp.
 * All Rights Reserved.
 *
@@ -54,8 +54,8 @@ typedef struct
   success_t success;    // success flag, 1 = yes, 0 = no, -1 = unset
   int arch;             // arch
   int syscall;          // syscall
-  int session_id;       // Login session id
-  int exit;             // Syscall exit code
+  uint32_t session_id;  // Login session id
+  long long exit;       // Syscall exit code
   int exit_is_set;      // Syscall exit code is valid
   char *hostname;       // remote hostname
   slist *filename;      // filename list
